+++ R/butcher.R
@@ -12,8 +12,10 @@ butcher::butcher
 #'
 #' @return Axed model_stack object.
 #'
-#' @examples
+#' @examplesIf rlang::is_installed("ranger") & rlang::is_installed("kernlab")
+#' 
 #' \donttest{
+#' 
 #' # build a regression model stack
 #' st <-
 #'   stacks() %>%
@@ -35,6 +37,7 @@ butcher::butcher
 #' 
 #' format(object.size(st))
 #' format(object.size(butchered_st))
+#' 
 #' }
 #' @name axe_model_stack
 NULL