+++ R/fit_members.R
@@ -20,8 +20,10 @@
 #' 
 #' @template note_example_data
 #' 
-#' @examples 
+#' @examplesIf rlang::is_installed("ranger") & rlang::is_installed("kernlab")
+#' 
 #' \donttest{
+#' 
 #' # see the "Example Data" section above for
 #' # clarification on the objects used in these examples!
 #' 
@@ -60,12 +62,15 @@
 #'   fit_members()
 #'   
 #' log_st
+#' 
 #' }
 #' 
 #' @family core verbs
 #' @export
 fit_members <- function(model_stack, ...) {
   check_model_stack(model_stack)
+  check_for_required_packages(model_stack)
+  check_empty_ellipses(...)
   
   dat <- model_stack[["train"]]
   
@@ -140,7 +145,7 @@ fit_member <- function(name, wflows, members_map, train_dat) {
   
   member_params <- 
     wflows[[member_row$name.x[1]]] %>%
-    dials::parameters() %>%
+    parsnip::extract_parameter_set_dials() %>%
     dplyr::pull(id)
   
   needs_finalizing <- length(member_params) != 0
@@ -156,7 +161,7 @@ fit_member <- function(name, wflows, members_map, train_dat) {
     
     new_member <- 
       tune::finalize_workflow(member_wf, member_metrics[,member_params]) %>%
-      generics::fit(data = train_dat)
+      parsnip::fit(data = train_dat)
   } else {
     member_model <-
       members_map %>%
@@ -165,7 +170,7 @@ fit_member <- function(name, wflows, members_map, train_dat) {
       dplyr::pull()
     
     new_member <-
-      generics::fit(wflows[[member_model[1]]], data = train_dat)
+      parsnip::fit(wflows[[member_model[1]]], data = train_dat)
   }
   
   new_member
@@ -217,4 +222,60 @@ check_model_stack <- function(model_stack) {
   } else {
     check_inherits(model_stack, "model_stack")
   }
+}
+
+# given a model stack, find the packages required to fit members and predict 
+# on new values, and error if any of them are not loaded
+check_for_required_packages <- function(x) {
+  # for dispatch to required_pkgs.workflow when model
+  # is loaded in a fresh environment
+  suppressPackageStartupMessages(requireNamespace("workflows"))
+  
+  pkgs <-
+    purrr::map(
+      x$model_defs,
+      parsnip::required_pkgs
+    ) %>%
+    unlist() %>%
+    unique()
+
+  installed <- purrr::map_lgl(
+    pkgs,
+    is_installed_
+  )
+  
+  if (any(!installed)) {
+    error_needs_install(pkgs, installed)
+  }
+
+  purrr::map(
+    pkgs,
+    ~suppressPackageStartupMessages(requireNamespace(.x, quietly = TRUE))
+  )
+  
+  invisible(TRUE)
+}
+
+# takes in a vector of package names and a logical vector giving
+# whether or not each is installed
+error_needs_install <- function(pkgs, installed) {
+  plural <- sum(!installed) != 1
+  
+  last_sep <- if (sum(!installed) == 2) {"` and `"} else {"`, and `"}
+  
+  need_install <- paste0(
+    "`",
+    glue::glue_collapse(pkgs[!installed], sep = "`, `", last = last_sep),
+    "`"
+  )
+  
+  glue_stop(
+    "The following package{if (plural) 's' else ''} ",
+    "need{if (plural) '' else 's'} to be installed before ",
+    "fitting members: {need_install}"
+  )
+}
+
+is_installed_ <- function(pkg) {
+  rlang::is_installed(pkg)
 }