+++ R/predict.R
@@ -15,8 +15,9 @@
 #'
 #' @template note_example_data
 #'
-#' @examples 
+#' @examplesIf rlang::is_installed("ranger") & rlang::is_installed("kernlab")
 #' \donttest{
+#' 
 #' # see the "Example Data" section above for
 #' # clarification on the data and tuning results
 #' # objects used in these examples!
@@ -62,6 +63,7 @@
 #'   type = "prob", 
 #'   members = TRUE
 #' )
+#' 
 #' }
 #'
 #' @importFrom stats predict