+++ tests/testthat/test_butcher.R
@@ -1,5 +1,3 @@
-context("butcher")
-
 if ((!on_cran()) || interactive()) {
   if (on_github()) {
     load(paste0(Sys.getenv("GITHUB_WORKSPACE"), "/tests/testthat/helper_data.Rda"))
@@ -7,6 +5,18 @@ if ((!on_cran()) || interactive()) {
     load(test_path("helper_data.Rda"))
   }
 }
+
+skip_if_not_installed("modeldata")
+library(modeldata)
+
+skip_if_not_installed("ranger")
+library(ranger)
+
+skip_if_not_installed("kernlab")
+library(kernlab)
+
+skip_if_not_installed("nnet")
+library(nnet)
 
 # Unit testing for the component members would duplicate unit testing
 # from the butcher package. Since the desired functionality to test
@@ -98,18 +108,7 @@ test_that("model_stack + butcher() works", {
 test_that("butchered model stack printing works", {
   skip_on_cran()
   
-  verify_output(
-    test_path("out/model_stack_reg_butcher.txt"),
-    {butcher(st_reg_1__)}
-  )
-  
-  verify_output(
-    test_path("out/model_stack_class_butcher.txt"),
-    {butcher(st_class_1__)}
-  )
-  
-  verify_output(
-    test_path("out/model_stack_log_butcher.txt"),
-    {butcher(st_log_1__)}
-  )
+  expect_snapshot(butcher(st_reg_1__))
+  expect_snapshot(butcher(st_class_1__))
+  expect_snapshot(butcher(st_log_1__))
 })