+++ tests/testthat/test_cran.R
@@ -1,8 +1,18 @@
-context("on cran")
-
 # many of the tests for core verbs take a significant amount of time to run.
 # skip most all of them on cran in favor of a minimal test that will at least
 # flag breakages, even if uninformatively.
+
+skip_if_not_installed("modeldata")
+library(modeldata)
+
+skip_if_not_installed("ranger")
+library(ranger)
+
+skip_if_not_installed("kernlab")
+library(kernlab)
+
+skip_if_not_installed("nnet")
+library(nnet)
 
 test_that("basic stacks pipeline works", {
   skip_if_not_installed("tibble")