+++ tests/testthat/test_expressions.R
@@ -1,8 +1,16 @@
-context("Prediction equations")
-
-## -----------------------------------------------------------------------------
-
 library(parsnip)
+
+skip_if_not_installed("modeldata")
+library(modeldata)
+
+skip_if_not_installed("ranger")
+library(ranger)
+
+skip_if_not_installed("kernlab")
+library(kernlab)
+
+skip_if_not_installed("nnet")
+library(nnet)
 
 ## -----------------------------------------------------------------------------