+++ tests/testthat/test_utils.R
@@ -1,5 +1,3 @@
-context("helpers")
-
 if ((!on_cran()) || interactive()) {
   if (on_github()) {
     load(paste0(Sys.getenv("GITHUB_WORKSPACE"), "/tests/testthat/helper_data.Rda"))
@@ -8,6 +6,17 @@ if ((!on_cran()) || interactive()) {
   }
 }
 
+skip_if_not_installed("modeldata")
+library(modeldata)
+
+skip_if_not_installed("ranger")
+library(ranger)
+
+skip_if_not_installed("kernlab")
+library(kernlab)
+
+skip_if_not_installed("nnet")
+library(nnet)
 
 test_that("object types relate as expected", {
   skip_on_cran()
@@ -57,4 +66,16 @@ test_that("misc. utilities work", {
   expect_error(glue_stop("howdy {yall}"), "howdy y'all")
   expect_warning(glue_warn("howdy {yall}"), "howdy y'all")
   expect_message(glue_message("howdy {yall}"), "howdy y'all")
+  
+  expect_warning(
+    check_empty_ellipses(yall),
+    "were passed: 'yall'"
+  )
+  
+  going <- "on"
+  
+  expect_warning(
+    check_empty_ellipses(hey = yall, what = "is", going),
+    "were passed: 'hey', 'what', 'going'"
+  )
 })