+++ NEWS.md
@@ -1,4 +1,22 @@
-# v0.2.2
+# stacks 0.2.3
+
+* Addressed deprecation warning in `add_candidates` (#99).
+* Improved clarity of warnings/errors related to failed hyperparameter 
+tuning and resample fitting (#110).
+* Reduced model stack object size and fixed bug where object size of model stack 
+inflated drastically after saving to file (#116). Also, regenerated example objects 
+with this change--saved model objects may need to be regenerated in order to 
+interface with newer versions of the package.
+* Introduced a `times` argument to `blend_predictions` that is passed on to
+`rsample::bootstraps` when fitting stacking coefficients. Reducing this
+argument from its default (`25`) greatly reduces the run time of 
+`blend_predictions` (#94).
+* The package will now load packages necessary for model fitting at 
+`fit_members()`, if available, and fail informatively if not (#118).
+* Fixed bug where meta-learner tuning would fail with outcome names and levels
+including the string `"class"` (#125).
+
+# stacks 0.2.2
 
 * Fixed errors arising from outcome levels that are not valid column 
   names in the multinomial classification setting. 
@@ -8,7 +26,7 @@
   model objects may need to be regenerated in order to build stacks combining
   models generated before and after this update.
 
-# v0.2.1
+# stacks 0.2.1
 
 * Updates for importing workflow sets that use the `add_variables()` 
   preprocessor. 
@@ -16,7 +34,7 @@
 * Performance and member plots now show the effect of multiple mixture values. 
 * Package diagrams now have alt text.
 
-# v0.2.0
+# stacks 0.2.0
 
 ## Breaking changes
 
@@ -56,6 +74,6 @@ be regenerated before predicting, plotting, printing, etc.
   more informative.
 * Various improvements to documentation.
 
-# v0.1.0
+# stacks 0.1.0
 
 Initial release!